Refactor db.ts, remove business logic from the class
