+++ plugin-server/src/utils/db/db.ts
@@ -584,120 +584,6 @@ export class DB {
         return client ? kafkaMessages : updatedPerson
     }
 
-    private getPropertyLastUpdatedAtDateTimeOrEpoch(person: Person, key: string): DateTime {
-        const lookup = person.properties_last_updated_at[key]
-        if (lookup) {
-            return DateTime.fromISO(lookup)
-        }
-        return DateTime.fromMillis(0)
-    }
-
-    private getPropertiesLastOperationOrSet(person: Person, key: string): PersonPropertyUpdateOperation {
-        if (!person.properties_last_operation || !(key in person.properties_last_operation)) {
-            return PersonPropertyUpdateOperation.Set
-        }
-        return person.properties_last_operation[key]
-    }
-
-    private updatePropertiesLastOperation(person: Person, key: string, value: PersonPropertyUpdateOperation) {
-        if (!person.properties_last_operation) {
-            person.properties_last_operation = {}
-        }
-        person.properties_last_operation[key] = value
-    }
-
-    private updatePersonPropertiesLocal(
-        person: Person,
-        properties: Properties,
-        propertiesOnce: Properties,
-        timestamp: DateTime
-    ): boolean {
-        // updates the person & returns true/false if anything was updated
-        let updatedSomething = false
-        Object.entries(propertiesOnce).forEach(([key, value]) => {
-            if (
-                !(key in person.properties) ||
-                (this.getPropertiesLastOperationOrSet(person, key) === PersonPropertyUpdateOperation.SetOnce &&
-                    this.getPropertyLastUpdatedAtDateTimeOrEpoch(person, key) > timestamp)
-            ) {
-                updatedSomething = true
-                person.properties[key] = value
-                this.updatePropertiesLastOperation(person, key, PersonPropertyUpdateOperation.SetOnce)
-                person.properties_last_updated_at[key] = timestamp.toISO()
-            }
-        })
-        // note that if the key appears twice we override it with set value here
-        Object.entries(properties).forEach(([key, value]) => {
-            if (
-                !(key in person.properties) ||
-                this.getPropertiesLastOperationOrSet(person, key) === PersonPropertyUpdateOperation.SetOnce ||
-                this.getPropertyLastUpdatedAtDateTimeOrEpoch(person, key) < timestamp
-            ) {
-                updatedSomething = true
-                person.properties[key] = value
-                this.updatePropertiesLastOperation(person, key, PersonPropertyUpdateOperation.Set)
-                person.properties_last_updated_at[key] = timestamp.toISO()
-            }
-        })
-        return updatedSomething
-    }
-
-    public async updatePersonProperties(
-        teamId: number,
-        distinctId: string,
-        properties: Properties,
-        propertiesOnce: Properties,
-        timestamp: DateTime
-    ): Promise<void> {
-        if (Object.keys(properties).length === 0 && Object.keys(propertiesOnce).length === 0) {
-            return
-        }
-
-        let person: Person | undefined
-        await this.postgresTransaction(async (client) => {
-            person = await this.fetchPerson(teamId, distinctId, client, true)
-            if (!person) {
-                throw new Error(
-                    `Could not find person with distinct id "${distinctId}" in team "${teamId}" to update props`
-                )
-            }
-
-            const shouldUpdate = this.updatePersonPropertiesLocal(person, properties, propertiesOnce, timestamp)
-            if (!shouldUpdate) {
-                return
-            }
-
-            const updateResult: QueryResult = await client.query(
-                `UPDATE posthog_person SET
-                    properties = $1,
-                    properties_last_updated_at = $2,
-                    properties_last_operation = $3,
-                    version = COALESCE(version, 0)::numeric + 1
-                WHERE id = $4
-                RETURNING version`,
-                [
-                    JSON.stringify(person.properties),
-                    JSON.stringify(person.properties_last_updated_at),
-                    JSON.stringify(person.properties_last_operation || {}),
-                    person.id,
-                ]
-            )
-            person.version = Number(updateResult.rows[0].version)
-        })
-
-        if (this.kafkaProducer && person) {
-            const kafkaMessage = generateKafkaPersonUpdateMessage(
-                timestamp,
-                person.properties,
-                person.team_id,
-                person.is_identified,
-                person.uuid,
-                person.version
-            )
-            await this.kafkaProducer.queueMessage(kafkaMessage)
-        }
-    }
-
     public async deletePerson(person: Person, client: PoolClient): Promise<ProducerRecord[]> {
         await client.query('DELETE FROM posthog_person WHERE team_id = $1 AND id = $2', [person.team_id, person.id])
         const kafkaMessages = []
